% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelfile.R
\name{excelfile}
\alias{excelfile}
\alias{print.excelfile}
\title{excelfile S3 Object}
\usage{
excelfile(file)

\method{print}{excelfile}(x, ...)
}
\arguments{
\item{file}{A character vector of length \code{1} representing a path to an
MS Excel file.}

\item{x}{An object of class 'excelfile'}

\item{...}{Other parameters}
}
\value{
An object of class \emph{excelfile}. This object is essentially a
list with the following elements:
\itemize{
  \item fileName: The full file name (including extension)
  \item fileSize: The size of the file (in bytes)
  \item created: The time the file was created
  \item modified: The time the file was last modified
  \item noOfSheets: The number of (active) spreadsheets
  \item sheets: Names of the spreadsheets
  \item data: A list of data frames, one for each spreadsheet
}
}
\description{
Constructs S3 objects of class 'excelfile', which contain properties
associated with .xls/.xlsx files.

S3 method for class 'excelFile'
}
