% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelFile.R
\name{locate_header}
\alias{locate_header}
\title{Find a Header}
\usage{
locate_header(df, hdr, quietly = TRUE)
}
\arguments{
\item{df}{An object of class \code{data.frame}}

\item{hdr}{A character vector containing the expected header values}

\item{quietly}{logical; whether to print out progress or not}
}
\value{
Returns an S3 object that is a marker to the row it occupies
}
\description{
Discover the likely location of a header in an Excel spreadsheet
}
\details{
Finds the row that likely contains the actual header
}
