% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelFile.R
\name{excelFile}
\alias{excelFile}
\title{Excel file objects}
\usage{
excelFile(file)
}
\arguments{
\item{file}{Path to an MS Excel file}
}
\description{
Constructs S3 objects of class 'excelFile', which contain
properties associated with .xls/.xlsx files
}
